# docs/swagger.yaml
basePath: /
definitions:
...
  models.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/models.PaginationResponse'
      request_id:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  models.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  models.ProgramRequest:
    properties:
      created_by:
        type: integer
      name:
        type: string
      priority:
        type: integer
      type_id:
        type: integer
      updated_by:
        type: integer
    type: object
...
  models.Type:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
host: helpdesk-dev.nopadol.com
info:
  contact: {}
  description: This is a sample swagger server in Go.
  title: Example API – Report Problem System
  version: "1.0"
paths:
...
  /api/v1/program/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific program
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get program details
      tags:
      - programs
  /api/v1/program/create:
    post:
      consumes:
      - application/json
      description: Create a new program
      parameters:
      - description: Program data
        in: body
        name: program
        required: true
        schema:
          $ref: '#/definitions/models.ProgramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create program
      tags:
      - programs
  /api/v1/program/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a program
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete program
      tags:
      - programs
  /api/v1/program/list:
    get:
      consumes:
      - application/json
      description: Get list of all programs with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List programs
      tags:
      - programs
  /api/v1/program/list/{query}:
    get:
      consumes:
      - application/json
      description: Search programs by query string with pagination
      parameters:
      - description: Search query (use 'all' for all programs)
        in: path
        name: query
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Search programs
      tags:
      - programs
  /api/v1/program/type/create:
    post:
      consumes:
      - application/json
      description: Add a new program type
      parameters:
      - description: Type data
        in: body
        name: type
        required: true
        schema:
          $ref: '#/definitions/models.Type'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Add program type
      tags:
      - programs
  /api/v1/program/type/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a program type
      parameters:
      - description: Type ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete program type
      tags:
      - programs
  /api/v1/program/type/list:
    get:
      consumes:
      - application/json
      description: Get all program types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get program types
      tags:
      - programs
  /api/v1/program/type/list/{query}:
    get:
      consumes:
      - application/json
      description: Search program types by query string
      parameters:
      - description: Search query (use 'all' for all types)
        in: path
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Search program types
      tags:
      - programs
  /api/v1/program/type/update/{id}:
    post:
      consumes:
      - application/json
      description: Update an existing program type
      parameters:
      - description: Type ID
        in: path
        name: id
        required: true
        type: string
      - description: Type data
        in: body
        name: type
        required: true
        schema:
          $ref: '#/definitions/models.Type'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update program type
      tags:
      - programs
  /api/v1/program/update/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing program
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: string
      - description: Program data
        in: body
        name: program
        required: true
        schema:
          $ref: '#/definitions/models.ProgramRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update program
      tags:
      - programs
...
swagger: "2.0"